start working on cli rework
