+++ helix-cli-old/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+name = "helix-cli"
+version = "1.0.110"
+edition = "2024"
+
+[dependencies]
+clap = { version = "4.5.40", features = ["derive"] }
+helixdb = { path = "../helixdb" }
+tempfile = "3.19.1"
+dirs = "6.0.0"
+serde = { version = "1.0.219", features = ["derive"] }
+sonic-rs = "0.5.1"
+uuid = { version = "1.16.0", features = ["v4"] }
+chrono = "0.4"
+libc = "0.2.172"
+tokio = { version = "1.45.0", features = ["full"] }
+open = "5.0.1"
+spinners = "4"
+indicatif = "0.17.11"
+toml = "0.8"
+reqwest = { version = "0.12", features = ["json"] }
+serde_json = "1.0"
+webbrowser = "1.0"
+tokio-tungstenite = "0.27.0"
+futures-util = "0.3.31"
+
+[target.'cfg(windows)'.dependencies]
+windows = { version = "0.61.1", features = [
+    "Win32_System_Threading",
+    "Win32_Foundation",
+] }
+
+[[bin]]
+name = "helix"
+path = "src/main.rs"
+
+[features]
+normal = ["helixdb/server"]
+ingestion = ["helixdb/full"]
+default = ["normal"]
